island: Move code from cc/environment/utils.py to cc/server_utils/file_utils.py
