+++ monkey/monkey_island/cc/server_utils/windows_permissions.py
